move chrome that goes around component from shared/ to chartContainer/
