+++ packages/charts/src/lib/chartContainer/ChartContainer.svelte
@@ -3,10 +3,10 @@
 </script>
 
 <script lang="ts">
-	import ExportBtns from '../shared/ExportBtns.svelte';
-	import Footer from '../shared/Footer.svelte';
-	import SubTitle from '../shared/SubTitle.svelte';
-	import Title from '../shared/Title.svelte';
+	import ExportBtns from './ExportBtns.svelte';
+	import Footer from './Footer.svelte';
+	import SubTitle from './SubTitle.svelte';
+	import Title from './Title.svelte';
 
 	// export props to popultae the container
 	export let title: string | null = null;
