+++ packages/charts/src/lib/chartContainer/Footer.svelte
