+++ packages/charts/src/lib/index.js
@@ -1,11 +1,11 @@
 export { default as ChartContainer } from './chartContainer/ChartContainer.svelte';
+export { default as ExportBtns } from './chartContainer/ExportBtns.svelte';
+export { default as Footer } from './chartContainer/Footer.svelte';
+export { default as SubTitle } from './chartContainer/SubTitle.svelte';
+export { default as Title } from './chartContainer/Title.svelte';
 
 export { default as ScatterPlot } from './scatterPlot/ScatterPlot.svelte';
 
 export { default as AxisX } from './shared/AxisX.svelte';
 export { default as AxisY } from './shared/AxisY.svelte';
-export { default as ExportBtns } from './shared/ExportBtns.svelte';
-export { default as Footer } from './shared/Footer.svelte';
 export { default as Scatter } from './shared/Scatter.svelte';
-export { default as SubTitle } from './shared/SubTitle.svelte';
-export { default as Title } from './shared/Title.svelte';
